include get_available_axis method re #43
