+++ mslice/models/alg_workspace_ops.py
@@ -1,6 +1,5 @@
 
-
-class Utility(object): #rename me
+class AlgWorkspaceOps(object):
 
     def _get_number_of_steps(self, axis):
         return int(max(1, (axis.end - axis.start)/axis.step))
@@ -27,13 +26,10 @@ class Utility(object): #rename me
         if axis.step is None:
             axis.step = (axis.end - axis.start)/100
 
-    # def get_available_axis(self, workspace, expected_ws_type):
-    #     dim_names = []
-    #     if isinstance(workspace, str):
-    #         workspace = self._workspace_provider.get_workspace_handle(workspace)
-    #     if isinstance(workspace, expected_ws_type):
-    #         for i in range(workspace.getNumDims()):
-    #             dim_names.append(workspace.getDimension(i).getName())
-    #     else:
-    #         raise TypeError('expected workspace of type {0}, got {1}'.format(expected_ws_type, type(workspace)))
-    #     return dim_names
+    def get_available_axis(self, workspace):
+        dim_names = []
+        if isinstance(workspace, str):
+            workspace = self._workspace_provider.get_workspace_handle(workspace)
+        for i in range(workspace.getNumDims()):
+            dim_names.append(workspace.getDimension(i).getName())
+        return dim_names
